tokens: fix bug w/ erroneous "time must be in future" error
By default we set the token to be 24 hours in the future, which is
required by our validation. The problem arises in that we set the
field to be 24-hours _in the local timezone_ in the future; even though
the validation runs in the database which requires that the time be
handled in UTC.

The fix is to use UTC in the controller as well, so that we get
"current UTC instant" + 24 hours; which doesn't trigger the validation.